Refactored to remove interface and replace with concrete class. Removed some blocks of repeated code.
